Moving composer components to styled components - Anchor + Button
