+++ packages/composer/composer/components/CloseButton.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Button from '../components/Button';
+import Button from '../components/styled/Button';
 import styles from './css/CloseButton.css';
 
 class closeButton extends React.Component {
