+++ packages/composer/composer/components/SuggestedMediaBox.jsx
@@ -9,7 +9,7 @@ import throttle from 'lodash.throttle';
 import { MediaTypes } from '@bufferapp/publish-constants';
 import ComposerActionCreators from '../action-creators/ComposerActionCreators';
 import AppActionCreators from '../action-creators/AppActionCreators';
-import Button from '../components/Button';
+import Button from '../components/styled/Button';
 import SuggestedMediaThumbnailInfo from './SuggestedMediaThumbnailInfo';
 import styles from './css/SuggestedMediaBox.css';
 import videoAttachmentThumbnailStyles from './css/VideoAttachmentThumbnail.css';
