--- packages/composer/composer/components/css/A.css
@@ -1,19 +1,0 @@
-/* Below are normalize.css rules for reliable base styles */
-
-/**
- * Remove the gray background color from active links in IE 10.
- */
-
-.a {
-  background-color: transparent;
-}
-
-/**
- * Improve readability of focused elements when they are also in an
- * active/hover state.
- */
-
-.a:active,
-.a:hover {
-  outline: 0;
-}
