+++ packages/composer/composer/components/styled/A.jsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const A = styled.a`
+  background-color: transparent;
+  :active,
+  :hover {
+    outline: 0;
+  }
+`;
+
+export default A;
