+++ packages/composer/composer/shared-components/modal/modal.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Button from '../../components/Button';
+import Button from '../../components/styled/Button';
 import styles from './css/modal.css';
 import ActionCreators from './actionCreators';
 import Store from './store';
